requirements.txt
firecrawl-py
tldextract
dotenv